Fix: locating file in a git submodule opens it in master git repository
Git submodules, are git repositories inside a git repository.

For exampole:

```
    /project
        .git/
            config
            modules/
                parts/
                    sub-project/
                    config
    parts/
        sub-project/
            .git <<<
                gitdir: ../.git/modules/parts/sub-project
            >>>

```

In a git-submodule the '.git' would be a file instead of directory.
It works like a symbolic link and points to location of .git folder.

Checkout https://git-scm.com/book/en/v2/Git-Tools-Submodules for more details
about git submodules.